home_screen_likes refactored and text replaced with text.rich
